requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12.1
Flask-SocketIO==2.8.6
gevent==1.2.1
gevent-websocket==0.10.1
greenlet==0.4.12
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
py-postgresql==1.2.1
pyparsing==2.2.0
python-dotenv==0.6.4
python-engineio==1.4.0
python-socketio==1.7.4
redis==2.10.5
requests==2.13.0
six==1.10.0
Werkzeug==0.12.1
